==> main/01_phyloseq_import.Rmd <==
---
title: "Phyloseq Import"
output: html_document
---


## Goal
Get cleaned data from Qiime2 into phyloseq.


```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

```


### Packages


```{r}

library(phyloseq)
library(qiime2R)
library(dplyr)
library(tibble)

```


## Data Import

Using the qiime2R package by jbisanz on Github. Reads in Qiime2 .qza files as is without needing to convert to .biom or .txt.

The data coming from Qiime2 has gone through quality control and had chimeras and bad quality reads removed. This is the full soil and root data set classified with the MaarjAM curated AMF database at 95%, 90%, and 80% ID at 90% query coverage each time. A final classification was done using AMF SSU data downloaded from NCBI GenBank at 80% ID, 90% query coverage.


```{r}

## This is more to access the qiime artifact data if desired, doesn't actually have to do with the phyloseq object yet

# Read in feature table
asvs <- read_qza("../data/merged-glom-only-final-no-negs.qza")

# Read in metadata
metadata <- read.table(file="../data/haka-metadata.tsv", sep="\t", header=TRUE, stringsAsFactors = TRUE)

# tree
tree <- read_qza("../data/rooted-tree.qza")

# Read in taxonomy
taxonomy <- read_qza(file="../data/total-merged-classification.qza")


```


## Create phyloseq object

```{r}

unrarefied_physeq <- qza_to_phyloseq(
  features="../data/merged-glom-only-final-no-negs.qza", 
  tree="../data/rooted-tree.qza", 
  taxonomy="../data/total-merged-classification.qza", 
  metadata="../data/haka-metadata.tsv")


```


## Data Housekeeping

```{r}

# Remove all pilo samples, root and soil
sample_data(unrarefied_physeq) <- sample_data(unrarefied_physeq)[                                    (sample_data(unrarefied_physeq)$host!="Coprosma rhynchocarpa"),]

# Fix Rubus 'hawaiensis' spelling 
sample_data(unrarefied_physeq)$host <- sub(pattern = "Rubus hawaiiensis", 
                                       replacement = "Rubus hawaiensis", 
                                       x = sample_data(unrarefied_physeq)$host)


## Add NewSampleType Column to sample_data

# Make data frame from sample data
sd_df <- as.data.frame(as(sample_data(unrarefied_physeq), "data.frame"))

# Change the value in new column NewSampleType according to existing values for SampleType and host. All right hand side must be same type, so NA_character_ must be used, not just NA
sd_df <- sd_df %>% 
  rownames_to_column('sampleid') %>%
  
  mutate(NewSampleType = case_when(
  (SampleType=="roots" & host=="Acacia koa") ~ "A. koa",
  (SampleType=="roots" & host=="Cheirodendron trigynum") ~ "C. trigynum",
  (SampleType=="roots" & host=="Myrsine lessertiana") ~ "M. lessertiana",
  (SampleType=="roots" & host=="Rubus hawaiensis") ~ "R. hawaiensis",
  (SampleType=="roots" & host=="Metrosideros polymorpha") ~ "M. polymorpha",
  (SampleType=="roots" & host=="Grass") ~ "Grass",
  (SampleType=="soil") ~ "Soil",
  TRUE ~ NA_character_)) %>% column_to_rownames('sampleid')

# reassign this data frame with the new column to be the sample_data 
# in the physeq object
sample_data(unrarefied_physeq)$NewSampleType <- sample_data(sd_df)$NewSampleType

## test - remove only those ASVs with 0 reads - 1973 ASVs
# test_physeq <- prune_taxa(taxa_sums(unrarefied_physeq) > 0, unrarefied_physeq)

## Remove ASVs present with 1 or 0 reads - 1932 ASVs
unrarefied_physeq <- prune_taxa(taxa_sums(unrarefied_physeq) > 1, unrarefied_physeq)

## test - remove only those ASVs with 0 reads
test_physeq <- prune_taxa(taxa_sums(unrarefied_physeq) > 0, unrarefied_physeq)

# Change the taxa names to "ASV1, ASV2, etc." for working convenience
taxa_names(unrarefied_physeq) <- paste0("ASV", seq(ntaxa(unrarefied_physeq)))

saveRDS(unrarefied_physeq, "../intermediate_files/rds/unrarefied_physeq.rds")


unrarefied_otu_table <- as.data.frame(otu_table(unrarefied_physeq))

saveRDS(unrarefied_otu_table, "../intermediate_files/rds/unrarefied_otu_table.rds")

```



## Subset Unrarefied Data by Host/Soil

```{r}

# initialize list for physeq objects
unrarefied_subsets <- list()

# make and save vector of host names + soil
host_names <- c("A. koa",
                "C. trigynum",
                "Grass",
                "M. polymorpha",
                "M. lessertiana",
                "R. hawaiensis",
                "Soil")

saveRDS(host_names, "../intermediate_files/rds/host_names.rds")


for (a_host in host_names) {
  # subset by host or soil
  unrare_subset <- subset_samples(unrarefied_physeq, NewSampleType==a_host)
  # remove empty taxa
  unrare_subset <- prune_taxa(taxa_sums(unrare_subset) > 0, unrare_subset)
  unrarefied_subsets[[a_host]] <- unrare_subset
}

saveRDS(unrarefied_subsets, "../intermediate_files/rds/unrarefied_subsets.rds")



### Do the same subsetting process but make lists by habitat

ro_unrarefied <- list()

for (a_host in host_names) {
  ro <- subset_samples(unrarefied_subsets[[a_host]], 
                 grepl("RO", sample_names(unrarefied_subsets[[a_host]])))
  ro_unrarefied[[a_host]] <- prune_taxa(taxa_sums(ro) > 0, ro)
}

saveRDS(ro_unrarefied, "../intermediate_files/rds/ro_unrarefied.rds")



ak_unrarefied <- list()

for (a_host in host_names) {
  ak <- subset_samples(unrarefied_subsets[[a_host]], 
                 grepl("AK", sample_names(unrarefied_subsets[[a_host]])))
  ak_unrarefied[[a_host]] <- prune_taxa(taxa_sums(ak) > 0, ak)
}

saveRDS(ak_unrarefied, "../intermediate_files/rds/ak_unrarefied.rds")


```


## Prepare Taxonomy Table for Later

```{r}

# extract taxonomy table and convert to data frame
full_tax_table_df <- as.data.frame(tax_table(unrarefied_physeq))

# format taxonomy table
full_tax_table_df <- full_tax_table_df %>% rownames_to_column("ASV")

# where family = Paraglomerales, change to Glomeromycotina
# Paraglomerales is an order, not a family. Issue stems from NCBI taxonomy.
full_tax_table_df$Family <- gsub("Paraglomerales", "Glomeromycotina", full_tax_table_df$Family)

saveRDS(full_tax_table_df, "../intermediate_files/rds/full_tax_table_df.rds")

```
