==> main/02_asv_accumulation_curves.Rmd <==
---
title: "ASV Accumulation Curves"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Libraries

```{r}

library(iNEXT)
library(ggplot2)
library(ggpubr)
library(phyloseq)

```

## Read in necessary files

```{r}

### lists of physeq objects

# unrarefied
ak_unrarefied <- readRDS("../intermediate_files/rds/ak_unrarefied.rds")
ro_unrarefied <- readRDS("../intermediate_files/rds/ro_unrarefied.rds")

### a vector of host names + soil
host_names <- readRDS("../intermediate_files/rds/host_names.rds")

```


## Unrarefied Data Accumulation Curves

```{r}


# extract ASV abundance sums (rowSums of the otu table with ASVs as rows, samples as columns)
get_abun_unrare <- function (physeq, host_or_soil) {
  
  abun_values <- rowSums(otu_table(physeq[[host_or_soil]]))
  
}


# apply function to habitat lists of unrarefied physeq objects
unrare_ro_abun_list <- lapply(host_names,
                get_abun_unrare,
                physeq = ro_unrarefied)

unrare_ak_abun_list <- lapply(host_names,
                get_abun_unrare,
                physeq = ak_unrarefied)

# name list items by host/soil
# use full binomial names

full_host_names <- c("Acacia koa",
                     "Cheirodendron trigynum",
                     "Grass",
                     "Metrosideros polymorpha",
                     "Myrsine lessertiana",
                     "Rubus hawaiensis",
                     "Soil")

names(unrare_ro_abun_list) <- full_host_names
names(unrare_ak_abun_list) <- full_host_names


# RO accumulation curve generation
unrare_ro_host_accum <- iNEXT(unrare_ro_abun_list, nboot = 200)

#plotdf <- fortify.iNEXT(unrare_ro_host_accum, type = 1)


unrare_ro_host_accum_curve <- ggiNEXT(unrare_ro_host_accum) +
  theme_classic() +
  ggtitle("Remnant Forest") +
  theme(plot.title = element_text(hjust = 0.5,
                                  size = 16,
                                  face = "bold",
                                  margin = margin(0,0,20,0)),
        axis.title.x = element_text(vjust=-2),
        axis.title.y = element_text(vjust=2)) +
  scale_shape_manual(values = c(19,19,19,19,19,19,19)) +
  scale_color_brewer(palette = "Set2") +
  scale_fill_brewer(palette = "Set2") +
  theme(plot.margin = margin(20,0,20,5)) +
  scale_y_continuous(limits=c(0,600), breaks=seq(0,600, by = 200)) +
  xlab("Number of sequences") +
  ylab("ASV richness")


# AK accumulation curve generation
unrare_ak_host_accum <- iNEXT(unrare_ak_abun_list, nboot = 200)

unrare_ak_host_accum_curve <- ggiNEXT(unrare_ak_host_accum) +
  theme_classic() +
  ggtitle("Restored Forest") +
  theme(plot.title = element_text(hjust = 0.5,
                                  size = 16,
                                  face = "bold",
                                  margin = margin(0,0,20,0)),
        axis.title.x = element_text(vjust=-2),
        axis.title.y = element_text(vjust=2)) +
  scale_shape_manual(values = c(19,19,19,19,19,19,19)) +
  scale_color_brewer(palette = "Set2") +
  scale_fill_brewer(palette = "Set2") +
  theme(plot.margin = margin(20,0,20,5)) +
  scale_y_continuous(limits=c(0,600), breaks=seq(0,600, by = 200)) +
  scale_x_continuous(limits=c(0,1000000), breaks=seq(0,1000000, by = 200000)) +
  xlab("Number of sequences") +
  ylab("ASV richness")


two_accums <- ggarrange(unrare_ro_host_accum_curve, 
                unrare_ak_host_accum_curve, 
                ncol = 2, 
                common.legend = T, 
                legend = "bottom",
                labels = c("A)", "B)"),
                vjust=3) + theme(plot.margin = margin(0,10,10,10))

ggsave("../figures/Fig_S2_ASV_accumulation_curves.png", width=11.5, height=5.5)


```
