==> main/04_venn.Rmd <==
---
title: "Venn"
output: ''
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Euler plot

Proportional venn diagram. Takes a data frame or matrix of presence-absence data with ASVs as row names and sample types as col names

Making these to visualize the overlap between remnant and restored ASVs found in a given host!



## Libraries

```{r}

library(eulerr)
library(ggpubr)

```

## Read in necessary files

```{r}

# abundance data frames for each habitat with ASVs as rows and hosts/soil
# as columns
ro_barplot_df <- readRDS("../intermediate_files/rds/ro_barplot_df.rds")
ak_barplot_df <- readRDS("../intermediate_files/rds/ak_barplot_df.rds")

```


## RO vs AK

```{r}

# extract all ASV names in a habitat 

# column number is arbitrary, just want to keep rownames (ASV names)
ak_present_asv_names <- ak_barplot_df[1]
# make column for ASV names
ak_present_asv_names$ASV <- rownames(ak_present_asv_names)
# keep only that ASV name column
ak_present_asv_names <- ak_present_asv_names[2]

# same as above
ro_present_asv_names <- ro_barplot_df[1]
ro_present_asv_names$ASV <- rownames(ro_present_asv_names)
ro_present_asv_names <- ro_present_asv_names[2]


# merge the two columns of ASV names by rownames (ASV names)
present_asv_name_df <- merge(ro_present_asv_names, 
                            ak_present_asv_names, 
                            by="row.names", 
                            all=T)

# rownames becomes the first column that we no longer need
present_asv_name_df <- present_asv_name_df[,2:3]
# rename!
names(present_asv_name_df) <- c("Remnant Forest", "Restored Forest")

# make a copy for changing ASV names and NAs to TRUE/FALSE
present_asv_name_bool <- present_asv_name_df
  
# if the value is NA, print False, otherwise print True (if the ASV exists)
present_asv_name_bool$`Restored Forest` <- ifelse(
  is.na(present_asv_name_bool$`Restored Forest`), F, T)
present_asv_name_bool$`Remnant Forest` <- ifelse(
  is.na(present_asv_name_bool$`Remnant Forest`), F, T)

# run euler function
all_asv_euler <- euler(present_asv_name_bool)
# make euler plot
all_euler_plot <- plot(all_asv_euler, 
                       fills=c("#ffadb6", "#b7e1f0"), 
                       quantities=TRUE,
                       edges=FALSE)



# export (paste in console)
pdf("figures/all_asv_venn.pdf", width=5, height=4)
all_euler_plot
dev.off()


```
## Soil vs Roots

```{r}

# root

ro_root_asv_names <- ro_barplot_df[1:6]

ro_root_asv_names <- ro_root_asv_names[rowSums(ro_root_asv_names) > 0,]

ro_root_asv_names <- ro_root_asv_names[1]


ak_root_asv_names <- ak_barplot_df[1:6]

ak_root_asv_names <- ak_root_asv_names[rowSums(ak_root_asv_names) > 0,]

ak_root_asv_names <- ak_root_asv_names[1]



all_root_asv_names <- merge(ro_root_asv_names, 
                            ak_root_asv_names, 
                            by="row.names",
                            all=T)

row.names(all_root_asv_names) <- all_root_asv_names$Row.names

all_root_asv_names <- all_root_asv_names[1]

names(all_root_asv_names)[1] <- "Root"



# soil

ro_soil_asv_names <- ro_barplot_df[7]

ro_soil_asv_names$ASV <- row.names(ro_soil_asv_names)

ro_soil_asv_names <- as.data.frame(ro_soil_asv_names[ro_soil_asv_names$Soil > 0,])

ro_soil_asv_names <- ro_soil_asv_names[2]


ak_soil_asv_names <- ak_barplot_df[7]

ak_soil_asv_names$ASV <- row.names(ak_soil_asv_names)

ak_soil_asv_names <- as.data.frame(ak_soil_asv_names[ak_soil_asv_names$Soil > 0,])

ak_soil_asv_names <- ak_soil_asv_names[2]



all_soil_asv_names <- merge(ro_soil_asv_names,
                            ak_soil_asv_names,
                            by="row.names",
                            all=T)

row.names(all_soil_asv_names) <- all_soil_asv_names$Row.names

all_soil_asv_names <- all_soil_asv_names[1]

names(all_soil_asv_names)[1] <- "Soil"



# merge

soil_vs_root_df <- merge(all_soil_asv_names,
                         all_root_asv_names,
                         by="row.names",
                         all=T)

soil_vs_root_df <- soil_vs_root_df[2:3]



# make a copy for changing ASV names and NAs to TRUE/FALSE
soil_root_bool <- soil_vs_root_df
  
# if the value is NA, print False, otherwise print True (if the ASV exists)
soil_root_bool$Soil <- ifelse(
  is.na(soil_root_bool$Soil), F, T)
soil_root_bool$Root <- ifelse(
  is.na(soil_root_bool$Root), F, T)

# run euler function
soil_root_euler <- euler(soil_root_bool)
# make euler plot
soil_root_plot <- plot(soil_root_euler, 
                       fills=c("#ffadb6", "#b7e1f0"), 
                       quantities=TRUE,
                       edges=FALSE)



## export
two_venn <- ggarrange(all_euler_plot, NULL, soil_root_plot, 
                      nrow=1, 
                      widths=c(1, 0.05, 1), 
                      labels = c("A)", "", "B)"))

ggsave("../figures/Fig_S3_venns.png", width=10, height=3.5)


```



## All ASVs by host

```{r}


# for seeing the habitat breakdown of all ASVs present by host/soil
present_asv_venn_list <- list()

for (a_host in host_names) {
  
  # extract unique ASV names for a host
  ak_present_asv_names <- ak_barplot_df[ak_barplot_df[a_host]!=0,
                         a_host,
                         drop=F]
  ak_present_asv_names$ASV <- rownames(ak_present_asv_names)
  ak_present_asv_names <- ak_present_asv_names[2]
  
  ro_present_asv_names <- ro_barplot_df[ro_barplot_df[a_host]!=0,
                         a_host,
                         drop=F]
  ro_present_asv_names$ASV <- rownames(ro_present_asv_names)
  ro_present_asv_names <- ro_present_asv_names[2]
  
  
  
  present_asv_name_df <- merge(ro_present_asv_names, 
                              ak_present_asv_names, 
                              by="row.names", 
                              all=T)
  
  present_asv_name_df <- present_asv_name_df[,2:3]
  names(present_asv_name_df) <- c("Remnant", "Restored")
  
  present_asv_name_bool <- present_asv_name_df
    
  present_asv_name_bool$Restored <- ifelse(
    is.na(present_asv_name_bool$Restored), F, T)
  present_asv_name_bool$Remnant <- ifelse(
    is.na(present_asv_name_bool$Remnant), F, T)
  
  present_asv_euler <- euler(present_asv_name_bool)
  present_euler_plot <- plot(present_asv_euler, fills=TRUE, quantities=TRUE)
  
  present_asv_venn_list[[a_host]] <- present_euler_plot
  
}

```




## Unique ASVs

```{r}

unique_venn_list <- list()

for (a_host in host_names) {
  
  # extract unique ASV names for a host (first column)
  ak_unique_asv_names <- ak_unique_asvs[[a_host]][1]
  rownames(ak_unique_asv_names) <- ak_unique_asv_names$ASV
  
  ro_unique_asv_names <- ro_unique_asvs[[a_host]][1]
  rownames(ro_unique_asv_names) <- ro_unique_asv_names$ASV
  
  
  unique_asv_name_df <- merge(ro_unique_asv_names, 
                              ak_unique_asv_names, 
                              by="row.names", 
                              all=T)
  
  unique_asv_name_df <- unique_asv_name_df[,2:3]
  names(unique_asv_name_df) <- c("Remnant", "Restored")
  
  unique_asv_name_bool <- unique_asv_name_df
    
  unique_asv_name_bool$Restored <- ifelse(
    is.na(unique_asv_name_bool$Restored), F, T)
  unique_asv_name_bool$Remnant <- ifelse(
    is.na(unique_asv_name_bool$Remnant), F, T)
  
  unique_asv_euler <- euler(unique_asv_name_bool)
  unique_euler_plot <- plot(unique_asv_euler, fills=TRUE, quantities=TRUE)
  
  unique_venn_list[[a_host]] <- unique_euler_plot
  
}


```



## Shared ASVs

```{r}

# we want a table per host of logical data with 2 columns, Restored and Remnant
# need to extract the shared ASV names and convert to boolean

# can also see 9_anova_hsd.Rmd for another method for extracting names

# subset shared ASVs. 
  # this code removes rows with any zero value
  # all() asks "are all values true?" and apply does this to all rows
  ak_shared_asv_names_df <- ak_barplot_df[apply(ak_barplot_df, 1, function(row) all(row!=0)),]
  ro_shared_asv_names_df <- ro_barplot_df[apply(ro_barplot_df, 1, function(row) all(row!=0)),]
  
  # make asv rownames into a column
  ak_shared_asv_names_df$ASVs <- row.names(ak_shared_asv_names_df)
  ak_shared_asv_names <- ak_shared_asv_names_df[8]
  
  
  ro_shared_asv_names_df$ASVs <- row.names(ro_shared_asv_names_df)
  ro_shared_asv_names <- ro_shared_asv_names_df[8]
  
  all_shared_asv_names <- list(ak_shared_asv_names, ro_shared_asv_names)
  all_shared_asv_names <- unique(Reduce(c, all_shared_asv_names))
  
  shared_asv_name_df <- merge(ro_shared_asv_names,
                              ak_shared_asv_names,
                              by="row.names",
                              all=T)
  
  shared_asv_name_df <- shared_asv_name_df[,2:3]
  names(shared_asv_name_df) <- c("Remnant", "Restored")
  
  shared_asv_name_bool <- shared_asv_name_df
    
  shared_asv_name_bool$Restored <- ifelse(is.na(shared_asv_name_bool$Restored), F, T)
  shared_asv_name_bool$Remnant <- ifelse(is.na(shared_asv_name_bool$Remnant), F, T)
  
  shared_asv_euler <- euler(shared_asv_name_bool)
  shared_euler <- plot(shared_asv_euler, fills=TRUE, quantities=TRUE)
  
```


# Checking

```{r}

# I know that 37% of ASVs are in soil from the venn
# but I wanted to confirm that 63% of the remaining ASVs
# were indeed only in hosts

# it wasn't working out because I was doing it by habitat, but there
# could have been ASVs that were in soil in one habitat and not in
# another habitat

# I needed to make all the host and soil columns from both habitats
# into one df and remove all the ones that were in soil in both
# of the habitats

v = merge(ak_barplot_df, ro_barplot_df, by = "row.names", all = T)

v = as.data.frame(v)

v = column_to_rownames(v, "Row.names")

v[is.na(v)] = 0


b = v[v$Soil.x==0 & v$Soil.y==0,]

```
